pyproject.toml
[tool.poetry]
name = "autofarm"
version = "0.1.0"
description = ""
authors = ["VitorDan <vitordan9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.26.1"
pandas = "^2.1.3"
plotly = "^5.18.0"
pymongo = "^4.6.0"
scikit-learn = "^1.3.2"
tensorflow = "^2.14.0"
tensorflow-io-gcs-filesystem = "^0.34.0"
python-binance = "^1.0.19"
ta-lib = "^0.4.28"
pandas-ta = "^0.3.14b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
